Modified random number generation for dgp.cpp
Now I use the Rcpp sugar function rnorm so I should be able to get
exactly the same draws as in the R version of dgp. I also created a
testing_code file to test this. I didn't get identical draws for x and
y so I went back and modified both the R and C++ versions to return the
errors epsilon, v, and eta. These were identical, so something is going
wrong in the parat of the code that generates the endogenous variables.
Perhaps a discrepancy from the zero indexing of C++? I need to figure
this out before proceeding...