First replication of Arellano & Bond (1991) Simulation Table 1
I added the summary statistics inline since the simulations take a
while to run. The results make qualitative sense but my estimators of
beta are biased downwards by a fair bit. There are a few possibilities:
(1) something is wrong with the results in the original paper, (2)
dgp.cpp code is somehow different from the DGP used in the paper, (3)
there's a mistake in my code for the estimator. I can try to rule out
(3) by going back over everything carefully as I rewrite in C++ to try
to speed things up. Should also go over the dgp function...